extract getPayload, simplify
