+++ test/get-payload.ts
@@ -1,0 +1,19 @@
+import { Buffer } from "node:buffer";
+import type { IncomingMessage } from "node:http";
+
+export function getPayload(request: IncomingMessage) {
+  return new Promise((resolve, reject) => {
+    const body = [] as Buffer[];
+    request.on("error", reject);
+    request.on("data", onData);
+    request.on("end", onEnd);
+
+    function onData(chunk: Buffer) {
+      body.push(chunk);
+    }
+
+    function onEnd() {
+      resolve(Buffer.concat(body).toString("utf8"));
+    }
+  });
+}
